# src/api/swagger.json
{
  "openapi": "3.0.0",
  "servers": [],
  "info": {
    "title": "",
    "description": "",
    "version": ""
  },
...
  "paths": {
...
    "/trainers": {
      "get": {
        "tags": [
          "Trainer"
        ],
        "summary": "List all trainers",
        "description": "Returns all trainers",
        "operationId": "FindAllTrainers",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Trainer"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Trainer"
        ],
        "summary": "Create new trainer",
        "description": "",
        "operationId": "CreateTrainer",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Trainer"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewTrainer"
        }
      }
    },
    "/trainers/{id}": {
      "get": {
        "tags": [
          "Trainer"
        ],
        "summary": "Get trainer",
        "description": "Returns the given trainer",
        "operationId": "FindTrainer",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of trainer to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Trainer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Trainer"
        ],
        "summary": "Update trainer",
        "description": "Updates the given trainer",
        "operationId": "UpdateTrainer",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of trainer to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Trainer"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewTrainer"
        }
      },
      "delete": {
        "tags": [
          "Trainer"
        ],
        "summary": "Delete trainer",
        "description": "Removes the given trainer",
        "operationId": "DeleteTrainer",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of trainer to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "requestBodies": {
...
      "NewTrainer": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewTrainer"
            }
          }
        },
        "description": "Trainer object that needs to be added to the database",
        "required": true
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "NewTrainer": {
        "type": "object",
        "properties": {
          "TName": {
            "type": "string"
          },
          "TShortName": {
            "type": "string"
          },
          "TPLEPartNumber": {
            "type": "string"
          },
          "TRevision": {
            "type": "string"
          },
          "TDescription": {
            "type": "string"
          },
          "TNetWeight": {
            "type": "integer"
          },
          "TGrossWeight": {
            "type": "integer"
          },
          "TEOL": {
            "type": "boolean"
          },
          "TTag": {
            "type": "string"
          },
          "modules": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
...
    }
  }
}